create graph interface
