+++ src/graph/graph_edge.rs
@@ -1,13 +1,14 @@
-use super::Graph;
+use super::graph_data::GraphData;
+use super::graph_impl::GraphImpl;
 
 #[allow(dead_code)]
-pub(crate) struct GraphEdge<'a> {
-    pub(super) graph: &'a Graph,
+pub(crate) struct GraphEdge<'a, Data: GraphData> {
+    pub(super) graph: &'a GraphImpl<Data>,
     pub(super) index: i64,
 }
 
 #[allow(dead_code)]
-impl<'a> GraphEdge<'a> {
+impl<'a, Data: GraphData> GraphEdge<'a, Data> {
     pub(crate) fn index(&self) -> i64 {
         self.index
     }
