+++ src/graph/graph_edge_iterator.rs
@@ -1,13 +1,14 @@
+use super::graph_data::GraphData;
 use super::graph_edge::GraphEdge;
-use super::Graph;
+use super::graph_impl::GraphImpl;
 
-pub(crate) struct GraphEdgeIterator<'a> {
-    pub(super) graph: &'a Graph,
+pub(crate) struct GraphEdgeIterator<'a, Data: GraphData> {
+    pub(super) graph: &'a GraphImpl<Data>,
     pub(super) index: i64,
 }
 
-impl<'a> Iterator for GraphEdgeIterator<'a> {
-    type Item = GraphEdge<'a>;
+impl<'a, Data: GraphData> Iterator for GraphEdgeIterator<'a, Data> {
+    type Item = GraphEdge<'a, Data>;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.index == 0 {
