+++ src/storage.rs
@@ -15,16 +15,17 @@ mod write_ahead_log;
 mod write_ahead_log_record;
 
 use crate::db_error::DbError;
-use storage_data::StorageData;
 use storage_record::StorageRecord;
 use storage_record_with_index::StorageRecordWithIndex;
 use write_ahead_log_record::WriteAheadLogRecord;
 
 #[allow(unused_imports)]
 pub(crate) use file_storage::FileStorage;
+pub(crate) use file_storage_data::FileStorageData;
 pub(crate) use serialize::Serialize;
 #[allow(unused_imports)]
 pub(crate) use stable_hash::StableHash;
+pub(crate) use storage_data::StorageData;
 #[allow(unused_imports)]
 pub(crate) use storage_hash_map::StorageHashMap;
 #[allow(unused_imports)]
